pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nova"
version = "2.0.0"
description = "NOVA — Next-gen Omnimodal Virtual Agent (Enterprise Edition)"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "websockets>=12.0",

    # AI / LLM
    "openai>=1.35",
    "anthropic>=0.28",

    # TTS
    "edge-tts>=6.1",

    # Vector store
    "qdrant-client>=1.9",

    # Config & Validation
    "tomli>=2.0; python_version < '3.11'",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",

    # Observability
    "prometheus-client>=0.20",
    "structlog>=24.2",

    # Security
    "PyJWT>=2.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "httpx",
    "locust>=2.29",
]
cosyvoice = [
    "torch>=2.3",
    "torchaudio>=2.3",
]
tracing = [
    "opentelemetry-api>=1.25",
    "opentelemetry-sdk>=1.25",
    "opentelemetry-exporter-otlp-proto-grpc>=1.25",
]
redis = [
    "redis>=5.0",
]
production = [
    "nova[tracing,redis]",
    "gunicorn>=22.0",
]

[project.scripts]
nova = "apps.nova_server.main:main"

[tool.hatch.build.targets.wheel]
packages = ["packages", "apps"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
